==> Mainproject/code/2.4. Data Available Summary.ipynb <==
# Summary Tables 

Summary tables below are summarized from "summary.csv", generated in "2.1.data_summary.R".
There are 25 species.

**Bold** = has data for a range of temperatures

# Coleoptera

|**No.**| **Order** |**Species**|**Development Time**|**Peak Fecundity Rate**|**Adult Mortality Rate**|**Juvenile Mortality Rate**|**Fecundity Loss Rate**| **Location**| **Remarks**
|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:----------:|---------:|
| | | |$a$|$b_{pk}$|$z$|$z_J$|$\kappa$|||
|1.|Coleoptera|*Anthonomus grandis*|**1.Development Rate(%), <br> 2.Development Time	(days)** | **Fecundity (Eggs/female/day)**|**1.Mortality Rate (%), <br> 2.Survivorship(% of live weevils per day)**|**Mortality Rate (%) [no time information, deleted]**|**Fecundity (Eggs/female/day)**|Lower Rio Grande Valley, Texas|There are two location information |
|2.|Coleoptera|*Stethorus punctillum*|**Development Rate(1/day)**|-|**Percentage Survival(%) [no time information,deleted]**|-|-|Quebec City, Canada|Development Rate has several information about stages (Eggs, 1st Instar, 	2nd Instar,3rd instar,4th instar, Pupae)|
|3.|Coleoptera|*Sitona discoideus*|	**Development Time (day / 1 individual)**|-|-|-|-|Qazvin province, Iran (collected); Takestan, Iran (measurement)|only have one stage information: egg|

# Diptera

|**No.**| **Order** |**Species**|**Development Time**|**Peak Fecundity Rate**|**Adult Mortality Rate**|**Juvenile Mortality Rate**|**Fecundity Loss Rate**| **Location**| **Remarks**|
|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:----------:|
| | | |$a$|$b_{pk}$|$z$|$z_J$|$\kappa$|||
|4.|Diptera|*Aedes aegypti*|-|**Oviposition Rate (Eggs per day per female)** |**1.Survival Time (days)** <br> **2.Mortality Rate (1/day)**|-|-|Marília,  São Paulo State, Brazil|The oviposition rate at different temperatures is considered to be maximum (bpk), as every temperature only have one oviposition rate.|
|5.|Diptera|*Aedes albopictus* | **Development Rate (1/day)**|**Oviposition Rate (eggs per female per cycle)**|**Percentage Survival(%)[no time information,deleted]**|**Percentage Survival(%)[no time information,deleted]**|-|Saint Pierre, France|-|
|6.|Diptera|*Aedes camptorhynchus*|**Development time (days)**|Fecundity (Eggs/7days) [only one temperature  information,deleted]|-|**Juvenile survival (proportion) [no time information,deleted]**|-|Breamlea flora reserver (near Geelong, Victoria, Australia)|Stages(larvae + pupae)for Development time|
|7.|Diptera|*Bactrocera correcta*| **Development Time (Days,Hours $\times$ individual $^{-1}$)**|-|-|-|-|Yuanjiang, Yunnan, China (collected); Kunming, Yunnan, China (measurement)|egg, pupae|
|8.|Diptera|*Aedes notoscriptus*|**Development time (Days)**|-|-|**Juvenile survival (Days)**|-|Inner metropolitan Adelaide, South Australia, Australia|Secondary Stressor: water type (Tap/Rain); stage: larvae|
|9.|Diptera|*Culex annulirostris*|**1.Development time(Days)**|**Fecundity Rate over 66 days (offspring $\times$ individual $^{-1}$ $\times$ day $^{-1}$)**|**1.Adult survival(1/day) [no temperature information,deleted]** <br> **2.Adult survival (female, bloodfed; male)(proportion)[no time information,deleted]** <br> **3.Adult longevity (male/female)**|**Juvenile survival (proportion) [no time information,deleted]**|-|Lab colony obtained from Plant Research Institute (Burnley, Victoria, Australia)| Adult survival/ longevity seperated into different categories, male/female|




# Hemiptera

| **No.**| **Order** |**Species**|**Development Time**|**Peak Fecundity Rate**|**Adult Mortality Rate**|**Juvenile Mortality Rate**|**Fecundity Loss Rate**|  **Location**|**Remarks**
|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:----------:|
| | | |$a$|$b_{pk}$|$z$|$z_J$|$\kappa$|||
|10.|Hemiptera|*Aphis gossypii*|**1.Development Rate (1/day), <br> 2.Development Time (Day/individual)** <br>|**Fecundity Rate (nymphs/female/day)** | **1.Longevity (day/individual)**, <br> **2.Survival Rate (Individuals/ day)**|**Mortality Rate [event/(individual *24 hours)]**|**Fecundity Rate (nymphs/female/day)** |Tehran, Iran;	Adana, Turkey| Mortality Rate (%) is about Nymph, which needs to be deleted; Mortality Rate [event/(individual *24 hours)]is about stage of juvenile, which needs to be retained|
|11.|Hemiptera|*Corythucha ciliata*| **1.Development Time (Days),** <br> **2.Development Rate<br> (100 individuals/day)**|**Fecundity (Eggs/ female)**|**1.Longevity (Days), <br> 2.Percentage Survival (%) [no time information, deleted]**|-|-|Longwu Road, Shanghai, China|Fecundity need to find experimental period, missing per day|
|12.|Hemiptera|*Planococcus citri*|**1.Generation Time <br>(Day/individual), <br>2.Development time  (Day/individual)**|-|**Longevity (Day/individual)**|-|-|Mahallat, Iran|Longevity : male/felmale; male data was deleted|




# Hymenoptera

| **No.**| **Order** |**Species**|**Development Time**|**Peak Fecundity Rate**|**Adult Mortality Rate**|**Juvenile Mortality Rate**|**Fecundity Loss Rate**| **Location**|**Remarks**
|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|
| | | |$a$|$b_{pk}$|$z$|$z_J$|$\kappa$|||
|13.|Hymenoptera|*Euplectrus ronnai*|**Development Time <br>(days/individual)**|-|-|**Mortality Rate (Deaths/(individual*24 hours))** |-|Curitiba, Parana, Brazil| development time: larvae/pupae; Mortality Rate: juvenile
|14.|Hymenoptera|*Glyptapanteles muesebecki*|**Development Time <br>(days/individual)**|-|-|-|-|Brazil|Development time: larvae/pupae;add them up|
|15.|Hymenoptera|*Macrocentrus iridescens*|**Development Rate (Individual completing development * individual/24 hours)**|-|-|-|-|Fresno, California|stage: juvenile|
|16.|Hymenoptera|*Telenomus chrysopae*|**Development Time <br>(days/individual)**|-|-|**Mortality Rate (Deaths/(individual*24 hours))**|-|Davis, Yolo County, California, USA (collected)& Ithaca, New York, USA (measurement); Ithaca, New York, USA |Both larvae|
|17.|Hymenoptera|*Telenomus isis*|**Development Rate (Individual completing development * (individual/ 24 hours))**|-|**Longevity (day/individual)**|**Mortality Rate (Deaths/(individual* 24 hours)**|-|Benin (collected); Nairobi, Kenya (measurement)|Development Time: juvenile; Longevity: adult (female); Mortality Rate: juvenile|
|18.|Hymenoptera|*Telenomus lobatus*|**Development Time <br>(days/individual)**|-|-|**Mortality Rate (Deaths/(individual* 24 hours)**|-|Manhattan, Riley County, Kansas, USA (collected); Ithaca, New York, USA (measurement)|Both larvae
|19.|Hymenoptera|*Theocolax elegans*|**Development Time <br>(days/individual)**|-|**Longevity (days/individual)**|-|-|Tsukuba, Ibaraki, Japan|Development Time:	juvenile (female) &	juvenile (male); Longevity: adult (female) & adult (male)|
|20.|Hymenoptera|*Trichogramma bruni*|**Development Time <br>(days/individual)**|-|-|-|-|Nairobi, Kenya|juvenile|
|21.|Hymenoptera|*Trichogramma sp. nr. Lutea*|**Development Time <br>(days/individual)**|-|-|-|-|Nairobi, Kenya|juvenile|
|22.|Hymenoptera|*Trichogramma sp. nr. Mwanzai*|**Development Time <br>(days/individual)**|-|-|-|-|Nairobi, Kenya|juvenile|

# Lepidoptera

| **No.**| **Order** |**Species**|**Development Time**|**Peak Fecundity Rate**|**Adult Mortality Rate**|**Juvenile Mortality Rate**|**Fecundity Loss Rate**| **Location**|**Remarks**|
|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|-------:|
| | | |$a$|$b_{pk}$|$z$|$z_J$|$\kappa$| ||
|23.| Lepidoptera|*Cydia pomonella*| **Development Time (days/individuals)**|-|-|-|-|Maragheh region, East Azerbaijan, Iran (collected); Tehran, Iran (measurement)|Stages of development time are egg/larvae/pupae|

# Psocoptera

| **No.** | **Order** |**Species**|**Development Time**|**Peak Fecundity Rate**|**Adult Mortality Rate**|**Juvenile Mortality Rate**|**Fecundity Loss Rate**| **Location**|**Remarks**|
|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|--------:|
| | | |$a$|$b_{pk}$|$z$|$z_J$|$\kappa$| ||
|24.| Psocoptera|*Lepinotus reticulatus*| **Development Time (days)**|-|**Survivorship(individuals/day)**|-|-|Manhattan, Kansas|Development time: 1st instar, 2nd instar, 3rd instar, 4th instar; Survivorship: proportion of nymphs surviving to adults|

# Prostigmata

| **No.** | **Order** |**Species**|**Development Time**|**Peak Fecundity Rate**|**Adult Mortality Rate**|**Juvenile Mortality Rate**|**Fecundity Loss Rate**| **Location**|**Remarks**|
|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|:--------:|
| | | |$a$|$b_{pk}$|$z$|$z_J$|$\kappa$| ||
|25.| Prostigmata|*Tetranychus mcdanieli*| **Development Rate (1/day)**|-|-|**Percentage Survival(%)**|-|Quebec City, Canada|Development rate: Egg, Larvae, 	Protonymph, Deutonymph; Percentage Survival: noun information|
